requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-dotenv==1.0.1
